Move detection box and closest face validations to face bounding box controller
